No CCM analytic solution
1) updated NoCsomeAnalyticSolution to calculate the carbonic anhydrase
unsaturated case so we no longer need CHconc_NoCsome.m
2) fixed bug in NoCsomeAnalyticalSolution.m:
The selection criteria for the whether to use the carbonic anhydrase
saturated or unsaturated case was selecting incorrectly. The CO2
concentration calculated by each were for all purposes equal. However,
due to matlab rounding error they were slightly different in exactly
the right way to cause the selection criteria to fail. I recoded a more
robust criteria which says that if the two cases are within 0.1% of
each other to count it as saturated.